% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-binomial_aggregation.R
\name{abinomial}
\alias{abinomial}
\title{Aggregate binomial data}
\usage{
abinomial(data)
}
\arguments{
\item{data}{Binomial data vector.}
}
\value{
Aggregated binomial data.
}
\description{
Aggregates binomial data using sufficient statistics for binomial samples.
For a sample \eqn{\boldsymbol{y} = \{y_1, \dots, y_n\}} with \eqn{y_i \sim \text{Bin}(n, p)}, the sample is aggregated into the sufficient statistic \cr \cr
\eqn{s = \sum_{i=1}^n y_i}.
}
\keyword{internal}
