% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPCtools.R
\name{add_cohort_column}
\alias{add_cohort_column}
\title{Add cohort column to data frame}
\usage{
add_cohort_column(df, age, period)
}
\arguments{
\item{df}{Data frame with age and period column.}

\item{age}{Name of age column.}

\item{period}{Name of period column.}
}
\value{
Data frame with additional column for birth cohorts
}
\description{
Adds a column for birth cohorts to a data frame, derived from specified age and period cohorts.
}
