% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPCtools.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Aggregate data}
\usage{
aggregate_data(
  grouped_data,
  gaussian = NULL,
  precision.scales = NULL,
  binomial = NULL,
  multinomial = NULL
)
}
\arguments{
\item{grouped_data}{Data frame, grouped by the group_data function.}

\item{gaussian}{Vector of names for the Gaussian columns to be aggregated.

\strong{Defaults to NULL (optional).}}

\item{precision.scales}{Matrix of precision scales for Gaussian columns.

\strong{Defaults to NULL (optional).}}

\item{binomial}{Vector of names for the binomial columns to be aggregated.

\strong{Defaults to NULL (optional).}}

\item{multinomial}{Vector of names for the multinomial columns to be aggregated.

\strong{Defaults to NULL (optional).}}
}
\value{
Aggregated data frame, aggregation performed within groups on specified columns.
}
\description{
Aggregates a grouped data frame into summarizing statistics within groups.
Aggregation is performed according to sufficient statistics for the specified distribution of the columns to be aggregated.
Possible distributions are: Gaussian, binomial, multinomial.
}
