% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreprocessing.R
\name{amultinomial}
\alias{amultinomial}
\title{Aggregate multinomial data}
\usage{
amultinomial(data, col_name, all_categories)
}
\arguments{
\item{data}{Multinomial data vector.}

\item{col_name}{Name of column with multinomial data.}

\item{all_categories}{Name of categories in multinomial distribution.}
}
\value{
Aggregated multinomial data.
}
\description{
Aggregates mutlinomial data into sufficient statistics for multinomial samples.
}
\keyword{internal}
