% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreprocessing.R
\name{as.APC.df}
\alias{as.APC.df}
\title{Add 1-indexed APC columns to data frame, and reorders rows: first by age, then by period.}
\usage{
as.APC.df(df, age = NULL, period = NULL)
}
\arguments{
\item{df}{Data frame with age and period column.}

\item{age}{Name of age column in data frame. \cr
\strong{Defaults to \code{"age"} if not specified.}}

\item{period}{Name of period column in data frame. \cr
\strong{Defaults to \code{"period"} if not specified.}}
}
\value{
Data frame with index columns for age, period and cohort, reordered by age and period.
}
\description{
Adds 1-indexed columns for age, period and cohort to a data frame with age and period columns.
\cr \code{age_index = age - min(age) + 1}
\cr \code{period_index = period - min(period) + 1}
\cr \code{cohort_index = period_index - age_index + max(age_index)}
}
