% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-binomial_aggregation.R
\name{binomial_aggregator}
\alias{binomial_aggregator}
\title{Aggregate binomial data}
\usage{
binomial_aggregator(df, col_name)
}
\arguments{
\item{df}{Data frame with binomial data column.}

\item{col_name}{Binomial data column.}
}
\value{
Aggregated binomial data column.
}
\description{
Aggregates binomial data into sufficient statistics for binomial samples. Uses \code{\link{abinomial}}.
}
\keyword{internal}
