% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-check_cols_exist.R
\name{check_cols_exist}
\alias{check_cols_exist}
\title{Check if a set of columns is missing from a data frame. For use in \code{\link{aggregate_df}}.}
\usage{
check_cols_exist(cols, df)
}
\arguments{
\item{cols}{String or vector of strings that is the name of the columns.}

\item{df}{Data frame}
}
\value{
Nothing. Casts an error message if any of the columns are missing.
}
\description{
Check if a set of columns is missing from a data frame. For use in \code{\link{aggregate_df}}.
}
\keyword{internal}
