% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreprocessing.R
\name{create_NA_structure}
\alias{create_NA_structure}
\title{Create NA structure across age, period and cohort groups}
\usage{
create_NA_structure(
  df,
  stratify_var,
  age_col = "age",
  period_col = "period",
  cohort_col = "cohort",
  include.random = FALSE
)
}
\arguments{
\item{df}{Data frame with age, period and cohort columns.}

\item{stratify_var}{Stratification variable to create an NA indicator structure for.}

\item{age_col}{Name of age column.

\strong{Defaults to "age" (optional).}}

\item{period_col}{Name of period column.

\strong{Defaults to "period" (optional).}}

\item{cohort_col}{Name of cohort column.

\strong{Defaults to "cohort" (optional).}}

\item{include.random}{Whether to include an index for each group, for random error terms in MAPC models.

** Defaults to FALSE (optional)**}
}
\value{
Data frame with NA structure across age-period-cohort groups and strata.
}
\description{
Creates NA structure indicating which strata each age, period and cohort group belongs to.
}
