% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreprocessing.R
\name{group_APC_data}
\alias{group_APC_data}
\title{Group age-period data}
\usage{
group_APC_data(df, age = NULL, period = NULL, stratify_by = NULL)
}
\arguments{
\item{df}{A data frame to group.}

\item{age}{Name of age column in data frame.}

\item{period}{Name of period column in data frame.}

\item{stratify_by}{Additional stratification variables.

\strong{Defaults to NULL (optional).}}
}
\value{
Data frame grouped by age and period, and additional factor levels for each strata.
}
\description{
Groups a data frame with an age column and a period column into groups of age and period. Optionally, additional stratification can be performed.
}
