% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-DataPreprocessing.R
\name{group_APC_df}
\alias{group_APC_df}
\title{Group age-period-cohort data}
\usage{
group_APC_df(df, age = NULL, period = NULL, stratify_by = NULL)
}
\arguments{
\item{df}{A data frame with age and period columns.}

\item{age}{Name of age column in data frame. \cr
\strong{Defaults to \code{"age"} if not specified.}}

\item{period}{Name of period column in data frame. \cr
\strong{Defaults to \code{"period"} if not specified.}}

\item{stratify_by}{Additional stratification variables. \cr
\strong{Defaults to NULL (optional).}}
}
\value{
Data frame grouped by age and period, and additional factor levels for each strata.
}
\description{
Groups a data frame with an age column and a period column into groups of age and period. Optionally, additional stratification can be performed.
}
