% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataPreprocessing.R
\name{group_data}
\alias{group_data}
\title{Group data frame}
\usage{
group_data(df, group_by = c(), stratify_by = c())
}
\arguments{
\item{df}{A data frame to group.}

\item{group_by}{Variables in data frame to group data by.}

\item{stratify_by}{Variables in data frame to stratify data by.}
}
\value{
Grouped data frame with factor levels for strata.
}
\description{
Groups data into groups specified by variables \code{group_by} and \code{stratify_by}.
}
