% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-DataPreprocessing.R
\name{group_df}
\alias{group_df}
\title{Group data frame}
\usage{
group_df(df, group_by = c(), stratify_by = c())
}
\arguments{
\item{df}{A data frame with grouping and/or stratification variables.}

\item{group_by}{Variables in data frame used to group the data. \cr
\strong{Defaults to empty list (optional).}}

\item{stratify_by}{Variables in data frame used to stratify the data. \cr
\strong{Defaults to empty list (optional).}}
}
\value{
Grouped data frame.
}
\description{
Groups data into groups specified by variables \code{group_by} and/or \code{stratify_by}.
Grouping is performed using using \code{dplyr}'s \link[dplyr:group_by]{group_by()} method, so that group manipulation method's from \code{dplyr} can be used on the resulting data frame.  \cr
At least one of \code{group_by} and \code{stratify_by} must be supplied by the user.
}
