% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-longest_run.R
\name{longest_run}
\alias{longest_run}
\title{Function for finding longest consecutive run of non-missing indices}
\usage{
longest_run(x)
}
\arguments{
\item{x}{A vector of indices for which to find the longest consecutive run of indices}
}
\value{
A named list, containing arguments:
\describe{
\item{\code{from}}{The first index of the longest run.}
\item{\code{to}}{The last index of the longest run.}
\item{\code{step}}{The stepsize of the longest run.}
}
}
\description{
Function for finding longest consecutive run of non-missing indices
}
\keyword{internal}
