% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-model_selection.R
\name{model_selection_criteria_table}
\alias{model_selection_criteria_table}
\title{Make a table of model fit scores}
\usage{
model_selection_criteria_table(
  dic_scores = NULL,
  waic_scores = NULL,
  log_scores = NULL
)
}
\arguments{
\item{dic_scores}{Named list of DIC scores. Names correspond to the model.}

\item{waic_scores}{Named list of WAIC scores. Names correspond to the model.}

\item{log_scores}{Named list of log-scores scores (derived from CPO scores). Names correspond to the model.}
}
\value{
A table summarizing the model fit scores, as a \code{\link[ggplot2]{ggplot}} object.
}
\description{
Generates a table summarizing the provided model fit scores.
}
\keyword{internal}
