% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-multinomial_aggregation.R
\name{multinomial_aggregator}
\alias{multinomial_aggregator}
\title{Aggregate multinomial data (No changes needed here)}
\usage{
multinomial_aggregator(df, col_name, all_categories = NULL)
}
\arguments{
\item{df}{Data frame with multinomial data column.}

\item{col_name}{Name of column with multinomial data.}

\item{all_categories}{Character vector of names of categories in multinomial distribution.}
}
\value{
Aggregated multinomial data column (a one-row data frame).
}
\description{
Aggregates multinomial data into sufficient statistics for multinomial samples. Uses \code{\link{amultinomial}}.
}
\keyword{internal}
