% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc-NHIS_bmi.R
\docType{data}
\name{NHIS_bmi}
\alias{NHIS_bmi}
\title{NHIS dataset with recordings of body mass index (BMI)}
\format{
A tibble with 671208 rows and 9 variables:
\describe{
\item{year}{Survey year}
\item{age}{Age of respondents}
\item{educaion}{Level of educational attainment of respondents, coded as:
\enumerate{
\item Less than high school / General Educational Development (GED)
\item High school diploma
\item Some college or Associate of Arts (AA) degree
\item Bachelor's degree or higher
}
}
\item{race}{Self-reported race of responsents, coded as:
\enumerate{
\item White
\item Black
\item Hispanic
\item Other
}
}
\item{female}{Sex of respondents, code as:
\enumerate{
\item Female
\item Male
}
}
\item{bmi}{Body mass index (BMI) of respondents, calculated as \eqn{BMI = \frac{weight\ [kg]}{height\ [m]^2}}.}
\item{strata}{Stratum identifier, for variance estimation.}
\item{psu}{Primary sampling unit identifier, for variance estimation}
\item{sampweight}{Survey weight, post-stratified.}
}
}
\source{
Data from the National Health Interview Survey (NHIS).
}
\usage{
NHIS_bmi
}
\description{
This dataset includes BMI recordings of civilians residing in the US, aged 25-85+. Observations are at the individual level and have been anonymized.
The data are form the survey years 1997-2023. For information about the NHIS, see \url{https://www.cdc.gov/nchs/nhis/index.html}.
Basic preprocessing steps (e.g., filtering by age, harmonizing variable names) were applied to standardize the data for use in this package.
}
\keyword{datasets}
