% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-number_of_groups.R
\name{number_of_groups}
\alias{number_of_groups}
\title{Count number of groups across a set of variables in a data frame}
\usage{
number_of_groups(df, group_by)
}
\arguments{
\item{df}{A data frame with grouping and/or stratification variables.}

\item{group_by}{Variables in data frame that defines a grouping of the data.}
}
\value{
Number of distinct groups and strata in the data frame.
}
\description{
Counts number of groups across specified grouping and stratification variables in a data frame.
At least one grouping or stratification variable must be provided.
}
\keyword{internal}
