% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-model_selection.R
\name{plot_model_selection_criteria}
\alias{plot_model_selection_criteria}
\title{Make a plot of model fit scores}
\usage{
plot_model_selection_criteria(scores, title = NULL, alphabetic = FALSE)
}
\arguments{
\item{scores}{Named list of model fit scores. Names correspond to the model.}

\item{title}{(Optional) Plot title; defaults to the name of \code{scores}.}

\item{alphabetic}{(Optional) TRUE/FALSE, indicating if models are to be sorted alphabetically along the horizontal axis. Defaults to \code{FALSE}.}
}
\value{
Plot summarizing the model fit scores, and ranking the models based on their scores.
}
\description{
Makes a plot summarizing the model fit scores of a set of models,
and ranks the models based on their score.
}
\details{
The function assumes lower scores are preferable, so the models with the lowest scores are ranked higher.
}
\keyword{internal}
