% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-resolve_column.R
\name{resolve_column}
\alias{resolve_column}
\title{Helper for evaluating column names, strings, or self-contained vectors}
\usage{
resolve_column(arg)
}
\arguments{
\item{arg}{A column name (unquoted or string) or a vector.}
}
\value{
A quosure, symbol, or literal vector depending on input.
}
\description{
Accepts:
\itemize{
\item Unquoted column names (tidy-eval style)
\item Quoted column names (as strings)
\item Self-contained vectors
}
}
\keyword{internal}
