% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-validate_lincombs_against_formula.R
\name{validate_lincombs_against_formula}
\alias{validate_lincombs_against_formula}
\title{Validate lincomb terms against an INLA formula}
\usage{
validate_lincombs_against_formula(lincombs, formula)
}
\arguments{
\item{lincombs}{A list of linear combinations (as generated by \code{generate_apc_lincombs()}).}

\item{formula}{The INLA model formula object (e.g., from \code{generate_MAPC_formula()}).}
}
\value{
Invisible TRUE if all terms match. Otherwise, stops with an informative error.
}
\description{
Checks that all variable names used in \code{inla.make.lincomb()} expressions
(inside a list of lincombs) are present in the provided INLA model formula.
}
