% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPCtools.R
\name{abinomial}
\alias{abinomial}
\title{Aggregate binomial data}
\usage{
abinomial(data)
}
\arguments{
\item{data}{Binomial data vector.}
}
\value{
Aggregated binomial data.
}
\description{
Aggregates binomial data into sufficient statistics for binomial samples.
}
\keyword{internal}
