% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-DataPreprocessing.R
\name{add_APC_by_match}
\alias{add_APC_by_match}
\title{Add 1-indexed age, period and cohort indices via match()}
\usage{
add_APC_by_match(df, age_name, period_name, age_order, period_order, M = 1)
}
\arguments{
\item{df}{Data frame}

\item{age_name}{Name of the age (or age-group) column (string).}

\item{period_name}{Name of the period (e.g. year) column (string).}

\item{age_order}{Character vector giving the desired ordering of age levels}

\item{period_order}{Vector (numeric or character) giving the desired ordering of periods}

\item{M}{Grid factor, defined as the ratio of age interval width to period interval width.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{         Data frame with new columns: age_index, period_index, cohort_index
}\if{html}{\out{</div>}}
}
\description{
Add 1-indexed age, period and cohort indices via match()
}
\keyword{internal}
