% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-DataPreprocessing.R
\name{add_cohort_column}
\alias{add_cohort_column}
\title{Add cohort column to data frame}
\usage{
add_cohort_column(data, age, period, cohort_name = "cohort")
}
\arguments{
\item{data}{Data frame with age and period column.}

\item{age}{Age column in \code{data}.}

\item{period}{Period column in \code{data}.}

\item{cohort_name}{Name of the cohort column to be created. Defaults to \code{"cohort"}.}
}
\value{
Data frame with additional column for birth cohorts added.
}
\description{
Adds a column for birth cohorts to a data frame, derived from specified age and period columns through the relation \code{cohort = period - age}.
}
