% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-DataPreprocessing.R
\name{add_cohort_index}
\alias{add_cohort_index}
\title{Add cohort column to data frame}
\usage{
add_cohort_index(df, age_index = NULL, period_index = NULL)
}
\arguments{
\item{df}{Data frame with age and period column.}

\item{age_index}{Name of age column in data frame. \cr
\strong{Defaults to \code{"age_index"} if not specified.}}

\item{period_index}{Name of period column in data frame. \cr
\strong{Defaults to \code{"period_index"} if not specified.}}
}
\value{
Data frame with additional column for cohort indices.
}
\description{
Adds a column for cohort indices to a data frame, derived from specified age and period index columns through the relationship \code{cohort index = period index - age index + max(age index)}.
}
