% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPCtools.R
\name{agaussian}
\alias{agaussian}
\title{Aggregate Gaussian data}
\usage{
agaussian(data, precision.scale = NULL)
}
\arguments{
\item{data}{Gaussian data vector.}

\item{precision.scale}{Scales for the precision of each Gaussian observation.

\strong{Defaults to NULL (optional).}}
}
\value{
Aggregated Gaussian data.
}
\description{
Aggregates Gaussian data into sufficient statistics for Gaussian samples.
}
\keyword{internal}
