% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-DataPreprocessing.R
\name{as.APC.df}
\alias{as.APC.df}
\title{Add 1-indexed APC columns to data frame, handling numeric or categorical age/period}
\usage{
as.APC.df(
  df,
  age = NULL,
  period = NULL,
  age_order = NULL,
  period_order = NULL,
  M = 1
)
}
\arguments{
\item{df}{A data.frame}

\item{age}{Name of the age column (string).
Defaults to "age" if both \code{age} and \code{period} are NULL.}

\item{period}{Name of the period column (string).
Defaults to "period" if both \code{age} and \code{period} are NULL.}

\item{age_order}{(Optional) Character vector giving the desired order of age levels.
If NULL and age is factor/character, uses \code{unique(sort(df[[age]]))}.}

\item{period_order}{(Optional) Vector (numeric or character) giving the desired order of periods.
If NULL and period is factor/character, uses \code{unique(sort(df[[period]]))}.}

\item{M}{Grid factor = (width of age intervals)/(width of period intervals).
Defaults to 1 (i.e. assuming equal sized age and period increments).}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{           The data frame with new columns \code{age_index}, \code{period_index}, \code{cohort_index},
                   and sorted by \code{(age_index, period_index)}.
}\if{html}{\out{</div>}}
}
\description{
Add 1-indexed APC columns to data frame, handling numeric or categorical age/period
}
