% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPCtools.R
\name{as.APC.df}
\alias{as.APC.df}
\title{Create APC data frame}
\usage{
as.APC.df(df, age, period)
}
\arguments{
\item{df}{Data frame with age and period columns}

\item{age}{Age column}

\item{period}{Period column}
}
\value{
Data frame with 1-indexed age, period and cohort columns.
}
\description{
Creates 1-indexed age, period and cohort columns indicating age-period-cohort groups.
}
