% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-ModelEstimation.R
\name{fit_all_MAPC}
\alias{fit_all_MAPC}
\title{Fit all configurations of MAPC models using INLA}
\usage{
fit_all_MAPC(
  data,
  response,
  family,
  stratify_by,
  reference_strata = NULL,
  age = "age",
  period = "period",
  apc_prior = "rw1",
  random = T,
  control.compute = list(config = TRUE, dic = TRUE, waic = TRUE, cpo = TRUE),
  binomial.n = NULL,
  poisson.offset = NULL,
  verbose = F
)
}
\arguments{
\item{data}{A \code{data.frame} containing the raw age, period, response, and stratification variables.}

\item{response}{A string naming the response (outcome) variable in \code{data}.}

\item{family}{Likelihood family of the response variable}

\item{stratify_by}{The column in \code{data} to use for stratification.}

\item{reference_strata}{Optional; value(s) of \code{strata_col} to set as the reference level(s). Defaults to \code{NULL}.}

\item{age}{The age column in \code{data}.}

\item{period}{The period column in \code{data}.}

\item{apc_prior}{A string specifying the prior for the age, period, and cohort effects (e.g. \code{"rw1"}, \code{"rw2"}). Defaults to \code{"rw1"}.}

\item{random}{Logical; if \code{TRUE}, include an overall random effect in the APC model. Defaults to \code{TRUE}.}

\item{control.compute}{A list of control.compute options passed to \code{INLA} (e.g. \code{config}, \code{dic}, \code{waic}, \code{cpo}). Defaults to \code{list(config=TRUE, dic=TRUE, waic=TRUE, cpo=TRUE)}.}

\item{binomial.n}{For the \code{family=binomial} likelihood. Either an integer giving the number of trials for the binomial response, or the name of the column containing the number of trials for each observation.}

\item{poisson.offset}{For the \code{family=poisson} likelihood. Either an integer giving the denominator for the Poisson count response, or the name of the column containing the denominator for each observation.}

\item{verbose}{This is argument is passed along to the \code{inla()} function that estimates the MAPC model. If \code{verbose=TRUE}, the \code{inla}-program runs in verbose mode, which can provide more informative error messages.}
}
\value{
A named list of \code{mapc} objects, one for each configuration of shared vs. stratum-specific time effects: APc, ApC, aPC, Apc, aPc, apC.
}
\description{
Fits all configurations of shared vs. stratum-specific time effects:
\describe{
\item{APc}{Shared age and period effects, stratum-specific cohort effects.}
\item{ApC}{Shared age and cohort effects, stratum-specific period effects.}
\item{aPC}{Shared period and cohort effects, stratum-specific age effects.}
\item{Apc}{Shared age effects, stratum-specific period and cohort effects.}
\item{aPc}{Shared period effects, stratum-specific age and cohort effects.}
\item{apC}{Shared cohort effects, stratum-specific age and period effects.}
}
Uses the \code{\link{fit_MAPC}} function.
}
\examples{
\dontrun{
fit <- fit_MAPC(
  data               = data,
  response           = deaths,
  strata_col         = region,
  reference_strata   = North,
  age                = age_group,
  period             = year,
  apc_prior          = "rw1",
  random             = TRUE
)
}

}
