% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPCtools.R
\name{gaussian_aggregator}
\alias{gaussian_aggregator}
\title{Aggregate Gaussian data}
\usage{
gaussian_aggregator(df, col_name, precision.scale = NULL)
}
\arguments{
\item{df}{Data frame with a Gaussian column.}

\item{col_name}{Name of Gaussian column.}

\item{precision.scale}{Scales for the precision of each Gaussian observation.

\strong{Defaults to NULL (optional).}}
}
\value{
Aggregated Gaussian column.
}
\description{
Aggregates Gaussian data into sufficient statistics for Gaussian samples.
}
\keyword{internal}
