% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-ModelEstimation.R
\name{generate_MAPC_formula}
\alias{generate_MAPC_formula}
\title{Generate MAPC formula for INLA}
\usage{
generate_MAPC_formula(
  df,
  APC_format,
  response,
  stratify_var,
  apc_prior = "rw1",
  intercept = F,
  random_term = TRUE,
  age = "age",
  period = "period",
  cohort = "cohort",
  age.stratified = NULL,
  period.stratified = NULL,
  cohort.stratified = NULL,
  extra.fixed = NULL,
  extra.random = NULL,
  random.models = NULL
)
}
\arguments{
\item{df}{Data frame for which MAPC models should be fit}

\item{APC_format}{A string where lower-case letters indicate stratum-specific time effects and upper-case letters indicate shared time effects.}

\item{response}{A string, name of the column in \code{df} that represents the response variable.}

\item{stratify_var}{Stratification variable. At least one time effect should be stratum-specific, and at least one should be shared.}

\item{apc_prior}{Which prior model to use for the time effects.\cr\strong{Defaults to "rw1" (optional).}}

\item{intercept}{Boolean, indicating if an overall intercept should be included in the formula.\cr\strong{Defaults to TRUE (optional).}}

\item{random_term}{Indicator, indicating if a random term should be included in the model.\cr\strong{Defaults to TRUE (optional).}}

\item{age}{Name of age column}

\item{period}{Name of period column}

\item{cohort}{Name of cohort column}

\item{age.stratified}{Base of age x strata column names. If not provided, defaults to \code{age_<strataify_var>}.\cr The default results in column names \code{age_<stratify_var>_i}, for each age index \eqn{i=1, \dots, I}.}

\item{period.stratified}{Base of period x strata column names. If not provided, defaults to \code{period_<stratify_var>}.\cr The default results in column names \code{period_<stratify_var>_j}, for each age index \eqn{j=1, \dots, J}.}

\item{cohort.stratified}{Base of cohort x strata column names. If not provided, defaults to \code{cohort_<stratify_var>}.\cr The default results in column names \code{cohort_<stratify_var>_k}, for each age index \eqn{k=1, \dots, K}.}

\item{extra.fixed}{Name of additional fixed effects.\cr\strong{Defaults to NULL (optional).}}

\item{extra.random}{Name of additional random effects.\cr\strong{Defaults to NULL (optional).}}

\item{random.models}{Models for random effects. Supported \code{INLA} models include \code{'iid'}, \code{'rw1'} and \code{'rw2'}.\cr\strong{Defaults to NULL (optional).}}
}
\value{
A formula object that can be passed to INLA to fit the desired MAPC model.
}
\description{
Based on APC-format, generate the proper formula to pass to INLA for fitting MAPC models.
}
