% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPCtools.R
\name{group_data}
\alias{group_data}
\title{Group data frame}
\usage{
group_data(df, group_by, stratify_by)
}
\arguments{
\item{df}{A data frame to group.}

\item{group_by}{Variables in data frame to group data by.}

\item{stratify_by}{Variables in data frame to stratify data by.}
}
\value{
Grouped data frame with factor levels for strata.
}
\description{
Groups data into groups indicated by specified grouping and stratification variables.
}
