% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPCtools.R
\name{multinomial_aggregator}
\alias{multinomial_aggregator}
\title{Aggregate multinomial data}
\usage{
multinomial_aggregator(df, col_name, all_categories)
}
\arguments{
\item{df}{Data frame with multinomial data column.}

\item{col_name}{Name of column with multinomial data.}

\item{all_categories}{Name of categories in multinomial distribution.}
}
\value{
Aggregated multinomial data column.
}
\description{
Aggregates binomial data into sufficient statistics for binomial samples.
}
\keyword{internal}
