% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPCtools.R
\name{number_of_groups}
\alias{number_of_groups}
\title{Count number of groups in data frame}
\usage{
number_of_groups(df, group_by, stratify_by)
}
\arguments{
\item{df}{Data frame.}

\item{group_by}{Columns to group data by.}

\item{stratify_by}{Columns to stratify data by.}
}
\value{
Number of distinct groups in the data frame.
}
\description{
Counts number of groups across specified grouping and stratification variables in a data frame
}
\keyword{internal}
